backend/requirements.txt
# ✅ FIXED: backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pymongo==4.6.0
pydantic[email]==2.5.0
torch==2.1.0
transformers==4.35.0
google-generativeai==0.3.0
google-api-python-client==2.108.0
Pillow==10.1.0
pytesseract==0.3.10
python-dotenv==1.0.0
python-dateutil==2.8.2
typing-extensions==4.8.0